+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinPlayerCollector.java
@@ -9,8 +9,6 @@ import android.os.Build;
 import com.bitmovin.analytics.BitmovinAnalytics;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
 import com.bitmovin.analytics.data.DeviceInformationProvider;
-import com.bitmovin.analytics.data.EventDataFactory;
-import com.bitmovin.analytics.data.UserIdProvider;
 import com.bitmovin.player.BitmovinPlayer;
 
 public class BitmovinPlayerCollector extends BitmovinAnalytics {
@@ -32,8 +30,8 @@ public class BitmovinPlayerCollector extends BitmovinAnalytics {
     }
 
     public void attachPlayer(BitmovinPlayer player) {
-        EventDataFactory factory = new EventDataFactory(this.bitmovinAnalyticsConfig, context, new DeviceInformationProvider(context, getUserAgent(context)), new UserIdProvider(context));
-        BitmovinSdkAdapter adapter = new BitmovinSdkAdapter(player, this.bitmovinAnalyticsConfig, factory,
+        DeviceInformationProvider deviceInformationProvider = new DeviceInformationProvider(context, getUserAgent(context));
+        BitmovinSdkAdapter adapter = new BitmovinSdkAdapter(player, this.bitmovinAnalyticsConfig, deviceInformationProvider,
                 this.playerStateMachine);
 
         this.attach(adapter);