+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerCollector.java
@@ -5,8 +5,6 @@ import android.content.Context;
 import com.bitmovin.analytics.BitmovinAnalytics;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
 import com.bitmovin.analytics.data.DeviceInformationProvider;
-import com.bitmovin.analytics.data.EventDataFactory;
-import com.bitmovin.analytics.data.UserIdProvider;
 import com.google.android.exoplayer2.ExoPlayer;
 
 public class ExoPlayerCollector extends BitmovinAnalytics {
@@ -34,8 +32,8 @@ public class ExoPlayerCollector extends BitmovinAnalytics {
     }
 
     public void attachPlayer(ExoPlayer player) {
-        EventDataFactory factory = new EventDataFactory(this.bitmovinAnalyticsConfig, context, new DeviceInformationProvider(context, ExoUtil.getUserAgent(context)), new UserIdProvider(context));
-        ExoPlayerAdapter adapter = new ExoPlayerAdapter(player, this.bitmovinAnalyticsConfig, factory,
+        DeviceInformationProvider deviceInformationProvider = new DeviceInformationProvider(context, ExoUtil.getUserAgent(context));
+        ExoPlayerAdapter adapter = new ExoPlayerAdapter(player, this.bitmovinAnalyticsConfig, deviceInformationProvider,
                 this.playerStateMachine);
 
         this.attach(adapter);