+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+### Added
+
+- ability to override `mpdUrl` and `m3u8Url` with the analytics configuration (AN-1919)
+
+### Changed
+
+- [internal] refactored `EventData` field population to use an `EventDataManipulationPipeline` (AN-1919)
+
 ## v1.20.0
 
 ### Added